Implement functionality to retreive, add and delete comments
