+++ api/posts.js
@@ -1,4 +1,3 @@
-const uuid = require('uuid').v4;
 const express = require('express');
 const mongoose = require('mongoose');
 const router = express.Router();
@@ -6,6 +5,7 @@ const router = express.Router();
 const Post = require('../models/Post.model');
 const User = require('../models/User.model');
 const Follower = require('../models/Follower.model');
+const Comment = require('../models/Comment.model');
 
 const auth = require('../middleware/auth.middleware');
 const upload = require('../middleware/imageUpload.middleware');
@@ -13,8 +13,6 @@ const upload = require('../middleware/imageUpload.middleware');
 const {
   newLikeNotification,
   removeLikeNotification,
-  newCommentNotification,
-  removeCommentNotification,
 } = require('../server-utils/notifications');
 
 // @route   POST /api/posts
@@ -36,7 +34,10 @@ router.post('/', auth, upload.array('images', 5), async (req, res) => {
       techStack: JSON.parse(techStack),
     };
     if (sourceCode) postObj.sourceCode = sourceCode;
+
     const post = await new Post(postObj).save();
+    await new Comment({ post: post._id, comments: [] }).save();
+
     res.status(201).json(post);
   } catch (err) {
     console.error(err);
@@ -125,9 +126,7 @@ router.get('/saves', auth, async (req, res) => {
 // @desc    Get a post by ID
 router.get('/:postId', async (req, res) => {
   try {
-    let post = await Post.findById(req.params.postId)
-      .populate('user')
-      .populate('comments.user');
+    let post = await Post.findById(req.params.postId).populate('user');
     if (!post) {
       return res.status(404).json({ msg: 'Post not found' });
     }
@@ -308,98 +307,6 @@ router.get('/like/:postId', async (req, res) => {
       return res.status(404).json({ msg: 'Post not found' });
     }
     res.status(200).json(post.likes);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ msg: 'Server error' });
-  }
-});
-
-// @route   POST /api/posts/comment/:postId
-// @desc    Add a new comment to post
-router.post('/comment/:postId', auth, async (req, res) => {
-  try {
-    if (req.body.text.length < 1) {
-      return res
-        .status(400)
-        .json({ msg: 'Comment must be atleast 1 character long' });
-    }
-
-    let post = await Post.findById(req.params.postId);
-    if (!post) {
-      return res.status(404).json({ msg: 'Post not found' });
-    }
-
-    const comment = {
-      _id: uuid(),
-      user: req.userId,
-      text: req.body.text,
-      date: Date.now(),
-    };
-
-    post.comments.unshift(comment);
-    post = await post.save();
-
-    post = await Post.populate(post, 'comments.user');
-
-    if (post.user.toString() !== req.userId) {
-      await newCommentNotification(
-        post.user.toString(),
-        req.userId,
-        req.params.postId,
-        comment._id,
-        req.body.text
-      );
-    }
-
-    res.status(201).json(post.comments);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ msg: 'Server error' });
-  }
-});
-
-// @route   DELETE /api/posts/comment/:postId/:commentId
-// @desc    Delete a comment
-router.delete('/comment/:postId/:commentId', auth, async (req, res) => {
-  try {
-    let post = await Post.findById(req.params.postId);
-    if (!post) {
-      return res.status(404).json({ msg: 'Post not found' });
-    }
-
-    const comment = post.comments.find(
-      (comment) => comment._id === req.params.commentId
-    );
-    if (!comment) {
-      return res.status(404).json({ msg: 'Comment not found' });
-    }
-
-    const user = await User.findById(req.userId);
-
-    if (comment.user.toString() === req.userId || user.role === 'root') {
-      const index = post.comments.findIndex(
-        (comment) => comment._id === req.params.commentId
-      );
-      post.comments.splice(index, 1);
-      post = await post.save();
-
-      post = await Post.populate(post, 'comments.user');
-
-      if (post.user.toString() !== req.userId) {
-        await removeCommentNotification(
-          post.user.toString(),
-          req.userId,
-          req.params.postId,
-          comment._id
-        );
-      }
-
-      res.status(200).json(post.comments);
-    } else {
-      res
-        .status(401)
-        .json({ msg: 'You are not authorized to delete this comment' });
-    }
   } catch (err) {
     console.error(err);
     res.status(500).json({ msg: 'Server error' });
